Separated dashboard into multiple files
