+++ src/seahawk/seahawk_deck/dash_styling/color_palette.py
@@ -1,0 +1,12 @@
+COLOR_CONSTS = {
+    "MAIN_WIN_BKG"      : "#0c0c0f",
+    "INACTIVE_TAB_BKG"  : "#141416",
+    "INACTIVE_TAB_BD"   : "#2d2938",
+    "ACTIVE_TAB_BKG"    : "#111113",
+    "SURFACE_DRK"       : "#18181b",
+    "SURFACE_BRIGHT"    : "#34343c",
+    "ACCENT_LIGHT"      : "#9172f8",
+    "ACCENT_DRK"        : "#5c4ba3",
+    "PRIMARY_TEXT"      : "#fff0f5",
+    "SECONDARY_TEXT"    : "#8f8c95",
+}
