+++ src/seahawk/seahawk_deck/dash_styling/numeric_data_widget.txt
